requirements.txt
fastapi
uvicorn
python-dotenv
requests
selenium
webdriver-manager
groq
bcrypt
loguru
secure
python-jose[cryptography]
python-multipart
pandas
openpyxl
httpx
